+++ componentes/Analytics/index.js
@@ -1,0 +1,2 @@
+export {Analytics} from "./Analytics"
+export { gtag } from "./lib/gtag"
