--- componentes/ButtonLogin/ButtonLogin.css
@@ -1,15 +1,0 @@
-.ButtonLogin{
-    background-color: #145C56;
-    color: #FFFFFF;
-    border-radius: 30px;
-    border: 1px solid #FFFFFF;
-    width: 100%;
-    height: 100%;
-    padding: 15px;
-    opacity: 1;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.ButtonLogin:hover {opacity: 0.6}
