--- componentes/FormConsultoria/FormConsultoria.module.css
@@ -1,53 +1,0 @@
-.containerFormConsultoria{
-    background: #E6ECF0;
-    height: 400px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.titleFormConsultoria{
-    font-family: 'Inter';
-    font-style: normal;
-    font-weight: 400;
-    font-size: 32px;
-    line-height: 130%;
-    text-align: center;
-    width: 851px;
-    margin-bottom: 98px;
-    margin-top: 123px;
-    color: #1F1F1F;
-}
-.mailFormConsultoria{
-    color:#1D856C;
-}
-.buttonFormConsultoria{
-    background-color: #1D856C;
-    border-radius: 100px;
-    color:#fff;
-    font-family: 'Inter';
-    font-style: normal;
-    font-weight: 400;
-    font-size: 32px;
-    line-height: 130%;
-    text-align: center;
-    font-size: 15px;
-    padding: 21px;
-    margin-bottom: 108px;
-}
-.buttonFormConsultoria:hover{
-    color:#fff;
-    opacity: 0.8;
-}
-@media screen and (max-width: 1023px) {
-    .containerFormConsultoria{
-        height: fit-content;
-        width: 100%;
-    }
-    
-    .titleFormConsultoria{
-        width: 80%;
-        font-size: 24px;
-    }
-}
