--- componentes/FormConsultoria/README.md
@@ -1,12 +1,0 @@
-# Formulario Consultoria
-
-```react
-responsive: true
-
----
-<FormConsultoria
-    title="Preencha o formulário abaixo para receber nossa consultoria ou, se quiser conversar sobre outro assunto, envie um e-mail para "
-    mail="contato@impulsogov.org."
-    link="/formulario"
-    />
-```
