--- componentes/FormConsultoria/index.js
@@ -1,1 +1,0 @@
-export { FormConsultoria } from "./Footer.jsx";
