--- componentes/ImagemFundo/ImagemFundo.module.css
@@ -1,181 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Inter:wght@400;500&display=swap");
-.containerImagemFundo{
-    height: fit-content;
-    width: 98%;
-}
-.containerImagemFundo a:hover{
-    color: #1D856C;
-}
-
-.containerImagemGradiente{
-    position: relative;
-}
-.imagemFundo{
-    width: 100%;
-    object-fit: cover;
-    border-radius: 0px 0px 64px 0px;
-}
-.gradienteImagemFundo{
-    display: flex;
-    flex-direction: column;
-    background-image: linear-gradient(to right, rgba(0,0,0,1), rgba(0, 0, 0, 0));
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    top: 0;
-    padding: 64px 0 81px 82px;
-}
-.containerCardsImagemFundo{
-    display: flex;
-}
-.chamadaImagemFundo{
-    color: #fff;
-    font-family: 'Inter';
-    font-style: normal;
-    font-weight: 400;
-    font-size: 32px;
-    width: 630px;
-}
-.chamadaColor{
-    color: #9DEECD;
-}
-.cardsImagemFundo{
-    display: flex;
-    flex-direction: row;
-    margin-top: 200px;
-}
-.cardImagemFundo{
-    color: #fff;
-    font-family: 'Inter';
-    font-style: normal;
-    font-weight: 400;
-    border-top: solid 1px rgba(255, 255, 255, 0.8);
-    width: 304px;
-    margin-right: 20px;
-    padding: 32px 10px 0 0;
-}
-.cardTitleImagemFundo{
-    font-size: 24px;
-    margin-bottom: 24px;
-}
-.cardBodyImagemFundo{
-    font-size: 16px;
-}
-.botaoImagemFundo{
-    margin-top: 48px;
-    margin-bottom: 100px;
-    border: 1px solid rgba(255, 255, 255, 0.25);
-    border-radius: 80px;
-    padding: 16px;
-    width: 337px;
-    height: 56px;
-    color:#fff;
-    font-family: 'Inter';
-    font-style: normal;
-    font-weight: 400;
-    font-size: 15px;
-    display: flex;
-    justify-content: center;
-}
-
-@media screen and (min-width: 1920px) {
-    .chamadaImagemFundo{
-        width: 880px;
-        font-size: 48px;
-    }
-    .cardsImagemFundo{
-        margin-top: 250px;
-    }
-    .cardImagemFundo{
-        width: 425px;
-        padding: 48px 10px 0 0;
-    }
-    .cardTitleImagemFundo{
-        font-size: 32px;
-        margin-bottom: 24px;
-    }
-    .cardBodyImagemFundo{
-        font-size: 20px;
-    }
-    .botaoImagemFundo{
-        display: flex;
-        align-items: center;
-        padding: 0;
-        width: 411px;
-        height: 64px;
-        margin-top: 78px;
-        font-size: 18px;
-    }
-}
-
-@media screen and (min-width: 1024px) and (max-width: 1439px) {
-    .gradienteImagemFundo{
-        width: 100%;
-        height: 100%;
-    }
-    .imagemFundo{
-        height: 800px;
-    }
-}
-
-@media screen and (min-width: 1440px) and (max-width: 1919px) {
-    .imagemFundo{
-        height: 850px;
-    }  
-}
-
-@media screen and (max-width: 1023px) {
-    .containerImagemFundo{
-        width: 100%;
-        margin: 0;
-    }
-        
-    .gradienteImagemFundo{
-        background-image: url("/imagem-fundo.png");
-        width: 100%;
-        height: fit-content;
-        position: relative;
-        padding: 15px;
-    }
-    
-    .imagemFundo{
-        height: 400px;
-        border-radius: 0;
-    }
-    .chamadaImagemFundo{
-        margin-top: 30px;
-        width: 95%;
-        font-size: 20px;
-    }
-    .cardsImagemFundo{
-        flex-direction: column;
-        margin-top: 40px;
-    }
-    .cardImagemFundo{
-        width: 300px;
-        margin-right: 20px;
-        padding: 32px 5px 0 0;
-        margin-bottom: 24px;
-    }
-    .cardTitleImagemFundo{
-        font-size: 18px;
-        margin-bottom: 24px;
-    }
-    .cardBodyImagemFundo{
-        font-size: 14px;
-    }
-    .botaoImagemFundo{
-        margin-left: 15px;
-        margin-top: 0px;
-        margin-bottom: 15px;
-        border-radius: 80px;
-        padding: 5px;
-        width: 260px;
-        height: 30px;
-        color:#fff;
-        font-size: 12px;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-    }
-}
