--- componentes/ImagemFundo/README.md
@@ -1,31 +1,0 @@
-# ImagemFundo
-
-```react
-responsive: true
-
----
-<ImagemFundo
-    chamada = "Com uma forma eficiente de monitorar e agir sobre indicadores, "
-    chamadacolor = "queremos ajudar gestores a transformar a saúde pública do país."
-    cards = {[
-        {
-            title : "Análises",
-            body : "Dados detalhados e recomendações pensadas para diferentes contextos de município. Em um só lugar."
-        },
-        {
-            title : "Consultoria",
-            body : "Profissionais capacitados e prontos para auxiliar, sem custo, os municípios na gestão do Previne Brasil."
-        },
-        {
-            title : "Previne Brasil",
-            body : "Todas as informações que o gestor precisa ter sobre o Programa, cálculo de metas, repasses financeiros e mais."
-        }
-    ]}
-    botao = {
-                {
-                    label : "Saiba mais sobre o impulso previne",
-                    url : "/impulsoprevine"
-                }
-    }
-    />
-```
