--- componentes/Imagens/ImagensFull.module.css
@@ -1,40 +1,0 @@
-.containerBanner1 {
-  display: flex;
-  height: 85%;
-}
-
-.imageContainerBanner1{
-  height: 600px;
-  width: 95%;
-  border-radius: 0px 0px 25px 0px;
-  object-fit: cover;
-}
-
-/* SMALL */
-@media screen and (max-width: 1023px) {
-  .imageContainerBanner1{
-    height: 300px;
-    border-radius: 0px 0px 15px 0px;
-}  
-}
-
-/* MEDIUM */
-@media screen and (min-width: 1024px)and (max-width: 1400px) {
-.containerBanner1 {
-  height: 70%;
-}
-.imageContainerBanner1{
-  height: 500px;
-  border-radius: 0px 0px 20px 0px;
-}
-}
-
-/* XLARGE */
-@media screen and (min-width: 1920px) {
-  .containerBanner1 {
-    height: 900px;
-  }
-  .imageContainerBanner1{
-    border-radius: 0px 0px 30px 0px;
-  }
-}
