--- componentes/Imagens/ImagensFull2.module.css
@@ -1,35 +1,0 @@
-.containerBanner1-2 {
-    display: flex;
-  }
-
-.imageContainerBanner2{
-  height: 600px;
-  width: 95%;
-  border-radius: 0px 25px 25px 0px;
-  object-fit: cover;
-}
-
-/* SMALL */
-@media screen and (max-width: 1023px) {
-  .imageContainerBanner2{
-    height: 300px;
-    border-radius: 0px 15px 15px 0px;
-    margin-bottom: 25px;
-  }  
-}
-
-/* MEDIUM */
-@media screen and (min-width: 1024px)and (max-width: 1400px) {
-  .imageContainerBanner2{
-    height: 500px;
-    border-radius: 0px 20px 20px 0px;
-  }
-}
-
-/* XLARGE */
-@media screen and (min-width: 1920px) {
-  .imageContainerBanner2{
-    height: 900px;
-    border-radius: 0px 30px 30px 0px;
-  }
-}
