--- componentes/Link/index.js
@@ -1,2 +1,0 @@
-export { default } from "./Link";
-export * from "./Link";
