--- componentes/NavBar/NavBar.module.css
@@ -1,228 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Inter:wght@40;500&display=swap");
-.themeColorIP{
-  background-color: #145C56;
-  color: #FFFFFF;
-}
-.themeWhite{
-  background-color: #FFFFFF;
-  color: #1F1F1F;
-}
-.container_navbar {
-    display: flex;
-    padding: 28px 40px;
-    top: 0;
-    margin: 0;
-    font-family: 'Inter';
-    font-style: normal;
-    font-weight: 40;
-    position: fixed;
-    z-index: 100;
-    width: 100%;
-  }
-  
-  .container_navbar a{
-    font-family: 'Inter';
-    font-style: normal;
-    font-weight: 40;
-  }
-  .container_navbar a:hover{
-    color: #1D856C;
-  }
-  .logoWrapper_navbar {
-    height: 24px;
-    width: 372px;
-  }
-  
-  .logo_navbar {
-    height: 24px;
-    width: 140px;
-  }
-  .dropdownContentNavBar{
-    display: none;
-    position: absolute;
-  } 
-  .iconDropdownNavBar{
-    position: relative;
-    height: 10px;
-  }
-  .iconDropdownContainerNavBar{
-    width: 0; 
-    height: 0; 
-    border-left: 25px solid transparent;
-    border-right: 25px solid transparent;
-    border-bottom: 25px solid #2EB380;
-  }
-  .dropdownNavBar:hover .dropdownContentNavBar{
-    display: block;
-  }
-  .dropdownItemNavBar{
-    color: #fff;
-    font-size: inherit;
-    padding: 10px 15px 10px 15px;
-  }
-  .dropdownItemNavBar:hover {
-    background-color: #145C56;
-  }
-  .dropdownItemNavBar a:hover {
-    color: #fff;
-  }
-
-  .itensDropdownNavBar{
-    background-color: #2EB380;
-    padding-bottom: 15px;
-    padding-top: 18px;
-  }
-  .links_navbar {
-    display: flex;
-    list-style: none;
-    margin: 0;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-  }
-  .linksNavBarMoblie{
-    display: none;
-  }
-  .link_navbar {
-    padding-right: 24px;
-    white-space: nowrap;
-  }
-  
-  .containerCidadeLogin_navbar {
-    align-items: center;
-    display: flex;
-    height: 100%;
-    margin-left: auto;
-    justify-content: center;
-  }
-  
-  .ButtonLoginBox_navbar {
-    height: 30px;
-    margin-right: 15px;
-    margin-left: 10px;
-  }
-
-  .citySelect_navbarColorIP{
-    background-color: #145C56;
-    color: #FFFFFF;
-    border: none;
-    text-align: center;
-    font-family: 'Inter';
-    font-style: normal;
-    font-weight: 40;
-  }
-  .citySelect_navbarWhite{
-    background-color: #FFF;
-    color: #1F1F1F;
-    border: none;
-    text-align: center;
-    font-family: 'Inter';
-    font-style: normal;
-    font-weight: 40;
-  }
-
-
-  .citySelect_navbar select {
-    display: none; /*hide original SELECT element:*/
-  }
-
-  .buttonMoblieColorIP {
-    display: none;
-  }
-  .buttonMoblieWhite{
-    display: none;
-  }
-  
-  @media screen and (max-width: 640px) {
-    .container_navbar {
-      padding: 16px 24px 16px 24px;
-      align-items: center;
-    }
-    .container_navbar a {
-      color: #1F1F1F;
-    }
-    .logoWrapper_navbar {
-      height: 21px;
-      width: 135px;
-      margin-right: auto;
-    }
-    .navBarIcon{
-      background-color: #FFFFFF;
-    }
-    .buttonMoblieColorIP{
-      display: block;
-      margin-left: auto;
-      border: 1px solid #FFF;
-      border-radius: 50%;
-      padding: 8px;
-    }
-    .buttonMoblieWhite{
-      display: block;
-      margin-left: auto;
-      border: 1px solid #1F1F1F;
-      border-radius: 50%;
-      padding: 8px;
-    }
-    .buttonMoblieWhite:hover{
-      cursor: pointer;
-    }
-    .buttonMoblieColorIP:hover{
-      cursor: pointer;
-    }
-
-    .linksNavBarMoblie{
-      background-color: #2EB380;
-      width: 100%;
-      padding-top: 25px;
-      padding-bottom: 25px;
-      text-align: left;
-      display: none;
-    }
-    .linksNavBarMoblieVisible{
-      display: block;
-    }
-    .linksNavBarMoblie a{
-      color: #fff;
-      font-family: 'Inter';
-      font-style: normal;
-      font-weight: 40;
-      font-size: 18px;
-      margin-right: 18px;
-      }
-    .links_navbar{
-      display: none;
-    }
-    .link_navbar {
-      text-align: right;
-      margin-bottom: 10px;
-      padding: 0;
-    }
-    .linkNavBarMoblie {
-      margin-bottom: 10px;
-    }
-
-    .subMenuMoblie{
-      display: flex;
-      flex-direction: column;
-      margin-top: 10px;
-      padding-top: 10px;
-    }
-    .containerCidadeLogin_navbar {
-      display: none;
-    }
-
-    .ButtonLoginBox_navbar {
-      height: 30px;
-      margin-right: 0px;
-      margin-top: 10px;
-      position: relative;
-      left: 0;
-    }
-  
-  }
-
-  @media screen and (max-width: 1024px) {
-    .logoWrapper_navbar{
-      margin-right: 15px;
-    }
-  }
