--- componentes/Parcerias/README.md
@@ -1,8 +1,0 @@
-# Parcerias
-
-```react
-responsive: true
-
----
-<Parcerias />
-```
