--- componentes/Parcerias/index.js
@@ -1,1 +1,0 @@
-export { Parcerias } from "./Parcerias.jsx";
