--- componentes/ParceriasTexto/README.md
@@ -1,11 +1,0 @@
-# Parcerias
-
-```react
-responsive: true
-
----
-<ParceriasTexto
-    text = "O Impulso Previne é uma realização da ImpulsoGov, uma organização brasileira de saúde pública, sem fins lucrativos, que trabalha ao lado de gestores municipais e estaduais para impulsionar o uso de dados e tecnologia no setor público."
-    label = "Junto à Impulso Gov, o projeto conta com uma rede de apoio institucional e financiadores:"
- />
-```
