--- componentes/ParceriasTexto/index.js
@@ -1,1 +1,0 @@
-export { ParceriasTexto } from "./Parcerias.jsx";
