--- componentes/Slider/README.md
@@ -1,34 +1,0 @@
-# Slider
-
-```react
-responsive: true
-
----
-<Slider 
-    titulo = {"Veja o que os outros municípios acharam"}
-    core = {[
-        {
-            titulo : "João Fulano",
-            subtitulo : "Coordenador de AP  |  Belo Horizonte - MG",
-            corpo : "Você preenche um formulário, nós entramos em contato em até três dias úteis e já marcamos a primeira conversa. Não precisa assinar nada, é sem burocracia. úteis e já marcamos a primeira conversa. Não precisa assinar nada, é sem burocracia."
-        },
-        {
-            titulo : "José Beltrano",
-            subtitulo : "Coordenador de AP  |  Santo André - SP",
-            corpo : "Você preenche um formulário, nós entramos em contato em até três dias úteis e já marcamos a primeira conversa. Não precisa assinar nada, é sem burocracia. úteis e já marcamos a primeira conversa. Não precisa assinar nada, é sem burocracia.Ajudou muito os andreenses"
-        },
-        {
-            titulo : "Antonio Sicrano",
-            subtitulo : "Coordenador de AP  |  Diadema - SP",
-            corpo : "Você preenche um formulário, nós entramos em contato em até três dias úteis e já marcamos a primeira conversa. Não precisa assinar nada, é sem burocracia. úteis e já marcamos a primeira conversa. Não precisa assinar nada, é sem burocracia.Ajudou muito os diademenses"
-        },
-        {
-            titulo : "Maria Fulana",
-            subtitulo : "Coordenador de AP  |  São Bernardo do Campo - SP",
-            corpo : "Você preenche um formulário, nós entramos em contato em até três dias úteis e já marcamos a primeira conversa. Não precisa assinar nada, é sem burocracia. úteis e já marcamos a primeira conversa. Não precisa assinar nada, é sem burocracia.Ajudou muito os são-bernardenses"
-        }
-    ]}
-    chamada = "TAMBÉM QUERO A CONSULTORIA"
-
-/>
-```
