--- componentes/Typography/index.js
@@ -1,2 +1,0 @@
-export { default } from "./Typography";
-export * from "./Typography";
