+++ package.json
@@ -9,9 +9,12 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@impulsogov/design-system": "^0.0.16",
     "bootstrap": "^5.1.3",
+    "core-js": "^3.23.5",
     "dompurify": "^2.3.8",
     "next": "12.1.6",
+    "next-transpile-modules": "^9.0.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-ga": "^3.3.1",
