+++ pages/indicadores/index.js
@@ -1,5 +1,5 @@
 import Layout from "../../componentes/Layout";
-import { IFrame } from "../../componentes/IFrame/IFrame.jsx";
+import { IFrame } from "@impulsogov/design-system";
 
 const Index = () => {
   return (
