+++ pages/manual-impulso-previne/index.js
@@ -1,6 +1,6 @@
 import Layout from "../../componentes/Layout";
 import { IFrame } from "../../componentes/IFrame/IFrame.jsx";
-import { HomeBanner } from "../../componentes/HomeBanner/HomeBanner"
+import { HomeBanner } from "@impulsogov/design-system"
 
 const Index = () => {
   return (
